[#28] Add styles to separate file and dont consider calendar login screen in testing
