+++ components/menu/calendar/Screens/FetchScreen/CheckboxGroup.js
@@ -7,6 +7,7 @@ import {
   View, FlatList, AsyncStorage, Text, TouchableOpacity, StyleSheet, Alert
 } from 'react-native';
 import { Button } from 'react-native-elements';
+import styles from './styles';
 
 
 export default class CheckboxGroup extends Component {
@@ -152,31 +153,3 @@ render() {
   );
 }
 }
-
-const styles = StyleSheet.create({
-  title: {
-    fontSize: 20,
-    fontWeight: 'bold',
-    marginVertical: '5%',
-    marginStart: '5%'
-  },
-  flatListContainer: {
-    marginTop: '5%',
-    height: '70%',
-    width: '80%',
-    marginLeft: '10%'
-  },
-  item: {
-    marginTop: 24,
-    padding: 30,
-    backgroundColor: 'pink',
-    fontSize: 24,
-    borderRadius: 5
-  },
-  button: {
-    position: 'relative',
-    width: '50%',
-    bottom: -10,
-    left: '25%'
-  }
-});
